// application/dto/auth/types.gen.ts
export type UserInfo = {
    /**
     * Id
     */
    id?: string;
    /**
     * Email
     */
    email?: string | null;
    /**
     * First Name
     */
    first_name?: string | null;
    /**
     * Last Name
     */
    last_name?: string | null;
    /**
     * Avatar Id
     */
    avatar_id?: string | null;
    /**
     * Username
     */
    username?: string | null;
    /**
     * Is Notifications Enabled
     */
    is_notifications_enabled?: boolean | null;
    /**
     * Language
     */
    language?: string | null;
    /**
     * Is Email Verified
     */
    is_email_verified?: boolean;
};