// application/dto/main/types.gen.ts
export type DashboardBase = {
    /**
     * Events Count
     */
    events_count?: number | null;
    /**
     * Tasks Count
     */
    tasks_count?: number | null;
    /**
     * Notes Count
     */
    notes_count?: number | null;
};